+++ maestro_worker_python/response.py
@@ -4,7 +4,7 @@ from pydantic import BaseModel
 
 
 class WorkerResponse(BaseModel):
-    billable_seconds: int | None
+    billable_seconds: float | None
     stats: dict[str, float]
     result: dict[str, Any]
